supertable/opann: total dedup order; unconditional sampler arithmetic
Two review fixes:

- The calibration dedup's sort ends in a cell tie-break. Once the
  fine-rank lookup became keyed by the surviving copy's cell, an
  arbitrary keep between equally-rankable equal-distance replicas
  made the stamped depth law depend on merge order - nondeterministic
  calibration across runs. The order is now total.

- The sampler's proportional-mapping products compute in u128. The
  u64 versions were in-range only by u32-boundedness arguments (the
  cell-local product sits exactly at the boundary: (2^32-1)^2 barely
  fits), and a future type widening would have silently distorted
  which rows get sampled. Unconditional correctness at zero cost on
  a once-per-pass path.

+++ src/supertable/writer.rs
@@ -7015,8 +7015,11 @@ pub(in crate::supertable) async fn recalibrate_probe_laws(
     // physical (tombstones included); each picked ordinal is mapped
     // proportionally onto the cell's live rows below.
     let sample_count = total_docs.min(opann::WIDTH_LAW_QUERY_SAMPLE as u64);
+    // u128 intermediates: the products are provably in-range today only
+    // because doc counts are u32-bounded — widening makes the sampler
+    // correct unconditionally, at zero cost on this once-per-pass path.
     let sample_ordinals: Vec<u64> = (0..sample_count)
-        .map(|i| i * total_docs / sample_count)
+        .map(|i| ((u128::from(i) * u128::from(total_docs)) / u128::from(sample_count)) as u64)
         .collect();
     let mut picks: BTreeMap<(usize, u32), Vec<u32>> = BTreeMap::new();
     let mut base = 0u64;
@@ -7061,7 +7064,7 @@ pub(in crate::supertable) async fn recalibrate_probe_laws(
             .max(1);
         let mut last_idx = usize::MAX;
         for &ordinal in ordinals {
-            let idx = ((u64::from(ordinal) * rows.len() as u64) / phys) as usize;
+            let idx = ((u128::from(ordinal) * rows.len() as u128) / u128::from(phys)) as usize;
             let idx = idx.min(rows.len() - 1);
             if idx == last_idx {
                 continue;
